requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
python-dotenv==1.0.0
google-generativeai==0.3.2
youtube-transcript-api==0.6.1
pytube==15.0.0
python-multipart==0.0.6
